Add files via upload
The underlying technology in this physics board is C++.There is a mathematics tool that helps one to perform basic mathematical operations as an aid to the physics numericals .There are quite descriptive diagrams like circuit diagram, lens and mirror diagrams etc. for a better explaination . Overall it can be used by mentors for teaching physics.